practice_microservice/requirements.txt
mysql-connector-python
python-decouple
python-dotenv